refactor: keep IP column change in migration only

+++ database/install.sql
@@ -566,7 +566,7 @@ CREATE TABLE `v2_user` (
                            `is_admin` tinyint(1) NOT NULL DEFAULT '0',
                            `last_login_at` int(11) DEFAULT NULL,
                            `is_staff` tinyint(1) NOT NULL DEFAULT '0',
-                           `last_login_ip` bigint(20) unsigned DEFAULT NULL,
+                           `last_login_ip` int(11) DEFAULT NULL,
                            `uuid` varchar(36) NOT NULL,
                            `group_id` int(11) DEFAULT NULL,
                            `plan_id` int(11) DEFAULT NULL,
